+++ sudoku.h
@@ -1,16 +1,18 @@
 #ifndef SUDOKU_H
 #define SUDOKU_H
 
+#include <stddef.h>
+
 typedef struct Sudoku
 {
-    int edge;
-    int *grid;
+    size_t edge;
+    size_t *grid;
 } Sudoku;
 
-int get(Sudoku *sudoku, int x, int y);
-void put(Sudoku *sudoku, int x, int y, int value);
+size_t get(Sudoku *sudoku, size_t x, size_t y);
+void put(Sudoku *sudoku, size_t x, size_t y, size_t value);
 void print_sudoku(Sudoku *sudoku);
-Sudoku *get_sudoku_from_file(char *pathname, int size);
+Sudoku *get_sudoku_from_file(char *pathname, size_t size);
 void delete_sudoku(Sudoku *sudoku);
 
 #endif
